requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
yt-dlp==2023.11.16
openai-whisper==20231117
opencv-python==4.8.1.78
moviepy==1.0.3
numpy==1.24.3
pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.1
aiofiles==23.2.1
ffmpeg-python==0.2.0
assemblyai==0.21.0
speech-recognition==3.10.0
pysrt==1.1.2
face-recognition==1.3.0
dlib==19.24.2
scikit-learn==1.3.2
scipy==1.11.4
librosa==0.10.1
webrtcvad==2.0.10